Add priority to projects

+++ db/migrate/20220308141833_add_priority_to_projects.rb
@@ -1,5 +1,5 @@
 class AddPriorityToProjects < ActiveRecord::Migration[7.0]
   def change
-    add_column :projects, :priority, :decimal, default: 5
+    add_column :projects, :priority, :integer, default: 5
   end
 end
